src/server/KVTaskClient.java
package server;

import exceptions.ManagerSaveException;

import java.io.IOException;
import java.net.URI;
import java.net.http.HttpClient;
import java.net.http.HttpRequest;
import java.net.http.HttpResponse;

import static java.net.HttpURLConnection.HTTP_OK;

public class KVTaskClient {

    private final String uri;
    private final String token;
    private final HttpClient httpClient;

    public KVTaskClient(String uri) throws IOException, InterruptedException {
        this.uri = uri;

        httpClient = HttpClient.newBuilder()
                .build();

        HttpRequest request = HttpRequest.newBuilder()
                .GET()
                .uri(URI.create(uri + "/register"))
                .build();
        HttpResponse<String> response = httpClient.send(request, HttpResponse.BodyHandlers.ofString());
        token = response.body();
    }

    public void put(String taskKey, String jsonData) throws IOException, InterruptedException {
        HttpRequest request = HttpRequest.newBuilder()
                .POST(HttpRequest.BodyPublishers.ofString(jsonData))
                .uri(URI.create(uri + "/save/" + taskKey + "?API_TOKEN=" + token))
                .build();
        HttpResponse<String> response = httpClient.send(request, HttpResponse.BodyHandlers.ofString());
        if (response.statusCode() != HTTP_OK) {
            throw new ManagerSaveException("Ошибка в методе KVTaskClient.put(), код ответа: " + response.statusCode());
        }
    }

    public String load(String taskKey) throws IOException, InterruptedException {
        HttpRequest request = HttpRequest.newBuilder()
                .GET()
                .uri(URI.create(uri + "/load/" + taskKey + "?API_TOKEN=" + token))
                .build();
        HttpResponse<String> response = httpClient.send(request, HttpResponse.BodyHandlers.ofString());
        return response.body();
    }
}
